repro EEVDF: add flag for skipping kernel build when the active version already matches
This was already the default behavior. With this flag, that default
can be changed, so that the kernel is always built even if the active
version seems to match (useful for testing the same kernel version
with different config options).

Note that the AUTOBUILD_KERNELS flag is still respected, as it
controls a different logic (automatically build the kernel, vs print a
message and stop, allowing the user to manually build/install the
correct kernel).

Signed-off-by: Cristian Prundeanu <cpru@amazon.com>